Last last debug

+++ main.cpp
@@ -74,7 +74,7 @@ for(k = 0; k<n; k++)
             {
                 if(i1 != k)
                 {
-                    int a = M[i1][k];
+                    float a = M[i1][k];
                     for(j=0;j<2*n;j++)
                     {
                         M[i1][j] = M[i1][j] - a*M[k][j];
